feat #9 move to ui root pacakge
