+++ presentation/src/main/java/com/dcs/presentation/ui/main/MainActivity.kt
@@ -5,6 +5,7 @@ import androidx.activity.ComponentActivity
 import androidx.activity.compose.setContent
 import androidx.activity.enableEdgeToEdge
 import com.dcs.presentation.core.theme.AllAboutMovieTheme
+import com.dcs.presentation.ui.MainApp
 import dagger.hilt.android.AndroidEntryPoint
 
 @AndroidEntryPoint
