+++ presentation/src/main/java/com/dcs/presentation/ui/setting/SettingScreen.kt
@@ -14,7 +14,7 @@ import androidx.compose.ui.unit.dp
 import androidx.hilt.navigation.compose.hiltViewModel
 import androidx.navigation.NavHostController
 import com.dcs.presentation.core.extensions.collectAsEffect
-import com.dcs.presentation.ui.main.Screen
+import com.dcs.presentation.ui.Screen
 
 @Composable
 fun SettingRoute(
